created CalculateCaloriesFromMacrosUseCase.kt

+++ composeApp/src/commonMain/kotlin/org/thelazybattley/macrotrack/domain/model/FoodMacros.kt
@@ -4,5 +4,5 @@ data class FoodMacros(
     val protein: Double,
     val carbs: Double,
     val fat: Double,
-    val calories: Double
+    val calories: Int
 )
